requirements.txt
torch
torchaudio
numpy
librosa
python-dotenv
soundfile
seaborn
tqdm
pandas